+++ src/services/SFTP.js
@@ -127,6 +127,15 @@ class ServiceSFTP extends ServiceBase {
 				.catch((error) => {
 					if (error.level === 'client-authentication') {
 						// Put a note in the log to remind users that a password can be set
+						if (client.options.privateKeyFile !== '') {
+							// If there was a keyfile yet we're at this point, it might have broken
+							channel.appendLocalisedInfo(
+								'key_file_not_working',
+								client.options.privateKeyFile,
+								client.options.username,
+							);
+						}
+
 						this.channel.appendLocalisedInfo('requesting_password');
 
 						// Offer to use a password
@@ -141,6 +150,9 @@ class ServiceSFTP extends ServiceBase {
 								}
 							});
 					} else {
+						// This likely ain't happening - let's just ditch the client and reject
+						this.destroyClient(client);
+
 						reject(error);
 					}
 				});
@@ -196,6 +208,13 @@ class ServiceSFTP extends ServiceBase {
 		});
 	}
 
+	destroyClient(client) {
+		if (client.sftp) {
+			client.sftp.end();
+			client.sftp = null;
+		}
+	}
+
 	/**
 	 * Returns a Promise eventually resolving to a new client instance, with the addition
 	 * of performing cleanup to ensure a maximum number of client instances exist.
@@ -208,7 +227,7 @@ class ServiceSFTP extends ServiceBase {
 			keys;
 
 		return new Promise((resolve) => {
-			if (this.clients[hash]) {
+			if (this.clients[hash] && this.clients[hash].sftp) {
 				// Return the existing client instance
 				this.clients[hash].lastUsed = date.getTime();
 
@@ -275,14 +294,17 @@ class ServiceSFTP extends ServiceBase {
 	 * @param {object} service
 	 */
 	getClientOptions(service) {
-		let options = {
-			host: service.host,
-			port: service.port,
-			username: service.username,
-			privateKey: this._getPrivateKey(),
-			keepaliveInterval: service.keepaliveInterval,
-			tryKeyboard: true
-		};
+		let sshKey = this._getPrivateKey(),
+			options = {
+				host: service.host,
+				port: service.port,
+				username: service.username,
+				privateKey: sshKey && sshKey.contents,
+				privateKeyFile: sshKey && sshKey.file,
+				passphrase: service.keyPassphrase || this.config.privateSSHKeyPassphrase,
+				keepaliveInterval: service.keepaliveInterval,
+				tryKeyboard: true
+			};
 
 		// Add a password, if set
 		if (service.password) {
@@ -556,7 +578,9 @@ class ServiceSFTP extends ServiceBase {
 			this.connect().then((client) => {
 				client.put(local, remote)
 					.then(resolve)
-					.catch(reject);
+					.catch((error) => {
+						reject(new Error(`${remote}: ${error.message}`));
+					});
 			});
 		});
 	}
@@ -602,7 +626,7 @@ class ServiceSFTP extends ServiceBase {
 							stream.pipe(write);
 						})
 						.catch((error) => {
-							throw(error.message);
+							throw(new Error(`${remote}: ${error.message}`));
 						});
 				});
 			});
@@ -806,11 +830,23 @@ class ServiceSFTP extends ServiceBase {
 	 * @param {string} file
 	 */
 	_getPrivateKey() {
-		let keyFile = this.config.service.privateKey || this.config.privateSSHKey,
-			homeDir, defaultKeyFiles, a;
+		let keyFile, homeDir, defaultKeyFiles, a;
+
+		keyFile = String(
+			this.config.service.privateKey ||
+			this.config.privateSSHKey ||
+			''
+		).trim();
 
 		if (fs.existsSync(keyFile)) {
-			return fs.readFileSync(keyFile, 'UTF-8');
+			return {
+				'file': keyFile,
+				'contents': fs.readFileSync(keyFile, 'UTF-8')
+			};
+		} else if (keyFile !== '') {
+			// File doesn't exist and wasn't empty
+			channel.appendLocalisedError('key_file_not_found', keyFile);
+			return false;
 		}
 
 		// Fall back to attempting to find by default
@@ -827,7 +863,10 @@ class ServiceSFTP extends ServiceBase {
 				this.config.service.privateKey = defaultKeyFiles[a];
 
 				// ... Then return
-				return fs.readFileSync(defaultKeyFiles[a], 'UTF-8');
+				return {
+					'file': defaultKeyFiles[a],
+					'contents': fs.readFileSync(defaultKeyFiles[a], 'UTF-8')
+				};
 			}
 		}
 	}
@@ -903,6 +942,7 @@ ServiceSFTP.defaults = {
 	username: '',
 	password: '',
 	privateKey: '',
+	keyPassphrase: '',
 	root: '/',
 	keepaliveInterval: 3000,
 	debug: false,