+++ CHANGELOG.md
@@ -5,3 +5,12 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## 0.1.0
 - Initial release
+
+## 0.1.1
+- Add support for key file paasphrase
+- Improve SFTP connection error details
+- Add multilingual support
+- Allow README to exist within extension space
+- Create views for watch and upload queue
+- General improvements to queue management & error handling
+- Improve diff to use a progress spinner when downloading files